miner/requirements.txt
pandas
gitpython
numpy
progress